+++ MarketManagementProject/MarketManagementProject/Product.cs
@@ -10,12 +10,12 @@ namespace MarketManagementProject
 
         public string ProductName { get; set; }
         public double ProductPrice { get; set; }
-        public int ProductCount { get; set; }
+        public double ProductCount { get; set; }
 
         public Category Category;
         public int ProductCode { get; set; }
 
-        public Product(string productname, double productprice, int productcount, Category category)
+        public Product(string productname, double productprice, double productcount, Category category)
         {
             count++;
             ProductName = productname;
@@ -24,5 +24,7 @@ namespace MarketManagementProject
             Category = category;
             ProductCode = count;
         }
+
+
     }
 }
